Fix Prometheus endpoint to return created alerts
Due to a bug we always returned an empty list even if alerts were
created.

This commit adds a new ops feature flag :prometheus_notify_max_alerts
which limits to amount of processable alerts in a single Prometheus
payload.

When enabled, it truncates incoming alerts if the amount of
alerts in the provided Prometheus payload exceeds 100.

Changelog: fixed