Update BG migration cop to match removed methods
Updated a cop that flags usages of calling scheduling methods on
`BackgroundMigrationWorker` based on the removal of some migration
helper methods, such as `migrate_async` and `bulk_migrate_in`.

Also changed the cop to not autocorrect offenses. It doesn't make sense
to do this anymore since we don't have a 1:1 mapping between migration
helpers and worker scheduling methods.