Fix workhorse make test
Make test should be able to run also on a clean clone of the
repository.

Do not generate the gitaly configuration if gitaly integration tests
are not needed.

If gitaly integration tests are requested, then make should be able to
generate all the necessary binaries and configuration files without
user interaction.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/349194