Fail remote mirror if LFS sync fails
Previously for a remote mirror update, `UpdateRemoteMirrorService` would
ignore the state of `Lfs::PushService#execute`, which prevented a
project maintainer from knowing that an actual problem existed with the
sync.

We now fail the mirror, while allowing the Git push to continue. This is
guarded behind the `remote_mirror_fail_on_lfs` feature flag, and
additional logging is added to track when this happens.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/340482

Changelog: changed