Fix flaky DAST profile schedule audit spec
In some conditions - the most trivial of which are when this spec is run
in isolation with a clean test database - the profile schedule and
profile created by this spec could have the same ID, in different
tables. The spec would then find the wrong audit event in the audit
events table, because it was looking up only by `target_id`, and fail as
a result.