Allow setting Sidekiq status only when opted in
This adds the `opt_in_sidekiq_status` feature flag. When this is
disabled, everything works as normal: we always set an entry in Redis
for a Sidekiq job's status.

When the flag is enabled, we will only perform the Redis SET command if
the job itself has opted in to tracking its status. It can do this on
the job class level:

    class SomeWorker
      sidekiq_options status_expiration: 1.hour
    end

Or on an individual call:

    SomeOtherWorker.with_status.perform_async

The aim here is to heavily reduce the number of SET commands we send to
the Sidekiq Redis, as most are unnecessary - most workers don't need to
have their status tracked.