Rename routes for project services
Most of the references to "services" have already
been renamed to "integrations" but the project services
have not been updated yet. This change takes a small step
towards updating them by renaming the route the user
sees to "integrations".

We also need to update all references to paths in HAML
views.

Changelog: changed