Change file input on upload license page to a dropzone
Substitutes a Rails-rendered file input with a Vue dropzone component on upload
license page. Now the user can use file picker window or drag-n-drop a file
onto the form.

Because the dropzone is used inside a Ruby form (not inside a Vue form), some
changes were made to upload_dropzone component to support this: Allow
upload_dropzone to use dynamic input name and to modify its input files
property when files are droppped on it.

Changelog: changed
EE: true