Add spec to validate we do not add more cross-database foreign keys
We are trying to remove all cross-database foreign keys in
https://gitlab.com/groups/gitlab-org/-/epics/7249 . Cross-database
foreign keys are foreign keys that span the future decomposed `ci` and
`main` databases. In an effort to avoid adding more while this work is
ongoing it will help to have a spec in place to prevent it. This spec
should also be kept after we finish that work to ensure nobody adds more
cross-database foreign keys in future.