Fix image comment note submission
This uses similar rounding logic as in [Design Management].

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/349389.

The crux of why rounding is necessary is that the image may be scaled to
fit in the UI, and so measurements may result in non-integer values.  In
addition, [MouseEvent.offsetX/Y are technically floats][1], not strictly
integers.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/offsetX

[Design Management]: https://gitlab.com/gitlab-org/gitlab/blob/73a6999553598919ea2db32bb3266e56cea03852/app/assets/javascripts/design_management/components/design_presentation.vue#L213-216

Changelog: fixed