Fix CI mirror table sync race condition
After consuming sync events, we need to remove those.
Before this commit, we used to use first id and last id range of events.
However, this caused a race condition. If an event is on-the-fly
transaction, and before and after events are tired to be consumed,
that event gets lost.