test: Only set HOME when spawning Gitaly
In the next commit we want to use GitalySetup.build_gitaly for local
testing. Prior to this change, HOME is set when calling `make`. But this
might break when Golang is installed through `asdf`.

This change moves setting HOME to where Gitaly is spawned only. We add a
little comment to explain this is done to make sure Git is not affected
by user configuration.