Reset the page to 1 when changing tabs
If a scope doesn't have the same number of pages, users can end up in a
state where they see nothing and the page appears broken.

By moving to the first page, we ensure we load the first set of
environments, which should pretty much always exist. If the count is 0,
then the page will be empty as expected.