Always autofill group slug
A fellow engineer found an issue in the combined registration page,
where the group slug wouldn't update after a failed validation.

I think typically this wasn't seen, because a slug validator would
increment the slug, but on some occasions if you were faster than
the validator, you could get past it.

This small change always updates the group slug as opposed to only when
the group name is empty.

part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/350221

Changelog: fixed