Use Bitbucket repo description over project description when importing
When importing a repo from Bitbucket Server the repo representation was
pulling the GitLab project description from the Bitbucket project not
the repo. A Bitbucket project is similar to a GitLab group, so the
Bitbucket project description isn't the necessarily the right thing to
use. Following the Bitbucket Server API docs https://docs.atlassian.com
/bitbucket-server/rest/7.19.2/bitbucket-rest.html#idp181 the repo
representation should just pull from the root of the raw response.

Changelog: fixed