Increase exclusive lease timeout
Because I have observed staging take longer than a minute
to complete this perform. A lease timeout which is too
short is against the point of the exclusive lease. The
timeout should ideally be exactly as long as the longest
possible execution time, but it's ok if we make the
timeout a bit longer than needed. I don't know for sure
if 30 minutes is long enough, but it is known to be an
improvement.

Changelog: changed
EE: true