Fix autoscroll to comments on issues
Add ResizeObserver to set scroll position if a comment is
selected. Still has a bunch of unwanted side effects like
jumping when window is resized, or when sidebars are collapsed
on smaller screens.

Remove ResizeObserver on user interaction
This avoids a bunch of bad and terrible bugs breaking scrolling
when user does something like toggle sidebar or click on text
editor. That otherwise triggers a resize

Changelog: fixed