Add `stack_trace_record` into transaction to observe `COMMIT`
This fixes usage of `after_commit` - a set of operations
that are executed after the object is committed. Before that
change the operations would be treated as being executed
in context of a transaction, even though the transaction
was already finished.

This introduces trace record object that receives callback
once the after commit/rollback hooks are to be executed.