Decrease urgency for GraphQL
We now have a separate SLI that measures query duration. So we are
still monitoring the performance of single GraphQL queries. But there
could be multiple queries batched in a single request, causing the
request to take longer.

Lowering the urgency to account for multiple queries will improve the
`rails_requests` SLI.