Show "Issues" menu when Jira issues is enabled
When GitLab issues is disabled but Jira issues enabled,
we still want to show the "Issues" menu to allow users
to access Jira issues.

This was a regression since in a previous MR, we moved
the Jira menu items from their own menu to the shared
"Issues" menu.

Changelog: fixed
EE: true