Improve Sidekiq jobs that use DB load balancing
Removing this FF changes the way we handle Sidekiq jobs that read from
DB replicas. The job will no longer be scheduled 1 second in the future
because this adds considerable overhead to Sidekiq and Redis.

The Sidekiq job would just sleep for at most 1 second if the replica is
not up to date.

Changelog: performance