Add user's preferred language to cache key of MR title partial
This is to fix a bug wherein the Edit button is being shown in a
different language from the current user's preferred language.

Before this fix, when the partial was cached and the language is
different, it'll stay that way until the cache is invalidated.

Now, the current user's preferred language will be taken into
consideration for the cache key.

Changelog: fixed