Support default templates for issues/MRs via .gitlab/ folders
For new issues or merge requests, if a template name of "Default"
(case-insensitive) exists in the template dropdown, it will be
selected. This does not override a template set via project
settings.

This allows default templates for issues and merge requests to
be provided via the .gitlab/issue_templates/ and
.gitlab/merge_request_templates/ folders.

Changelog: added