Revise and recombine instructions, remove img
Now that we've fixed the problem of having two workflows, we need to
completely eradicate any hint of there ever being two ways to do this.
The previous changes to the docs left some traces of the two methods,
so I've cleaned up the instructions while keeping the feature flag
historical information. Also removed an old, old image that we don't
need any more.