Full revision of Changes page
This page has been slated for a revamp for some time, but I needed
contiguous time to make it happen. It's less of an edit and more
of a full rewrite.

- Remove old images
- Reshoot a single, better image
- Rename subheadings
- Combine more subheadings into the top 'concept' section
- Rebuild the version history of every section that had a feature
  flag at some time in the past
- Rewrite pretty much everything
- Occasional good-natured sobbing in frustration

This page needs some help and love. Now that the info is collected
onto a single page, let's start shaping it into what it needs.