spec: Refactor detailed error tests to serialize messages
The UserMergeBranch RPC returns error details on some error conditions
which allow the caller to do more fine-grained error handling. The tests
which exercise this for the AccessCheckError is verifying that this
works by returning an error whose status details metadata is set to an
already-serialized string. This is hard to read and ultimately obscures
what kind of error details we want the server to return.

Convert the test to instead construct the serialized error details from
the protobuf definitions.