CTRT cleanup of merged results pipelines doc
REmove all the info that's duplicated on the MR pipelines
and merge trains page. Move merge train details to merge
trains page. Simplify language as much as possible, and
make sure the page follows current style guide for navigation
and content types. Also adds a header to the feature flags
admin page so that we can link to that section.