Fix long label overflow in the sidebar
When they overflow, menu items in the sidebar have their text that
continue a few pixels beyond the position it should stop to take into
account their margin, and if the menu item contains a badge, it is not
aligned as expected.

By setting `overflow: hidden`, both problems are adressed.  For a better
visual clue of the truncation, `text-overflow: ellipsis` is combined to
the this.

Changelog: fixed