Fix exception when epic has no author
Failure to assign the ghost author or delete epic when a user is
deleted with contributions has left some with no author. This
causes the Epic List to break as it tries to link to an author
that isn't there. This fixes the link_to_member method to handle
a missing author so at least this page will load.

Changelog: fixed