Update project export job status upon failure
When a project export is initiated, we create a `ProjectExportJob`
database record that tracks the progress of the Sidekiq job. If an
export fails, the record may cause the project export status to
indicate "in progress" or "regeneration in progress" until the
`StuckExportJobsWorker` comes along and updates the status as failed.

We can expedite this process by handling failures and marking the job
status appropriately.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/351155

Changelog: fixed