Sort vulnerability identifiers on ingestion to prevent Deadlock errors
It is possible to have multiple threads trying to ingest vulnerability
identifiers for the same project in different orders which can cause
multiple connections to wait for each other to release row locks to
proceed which introduces a deadlock error.

By sorting the identifiers by fingerprint before ingestion, we prevent
the risk of having multiple transactions to wait for each other.

Changelog: fixed
EE: true