Fix loose foreign key test to use right connection
Similar to what we did in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/75511 we need to
ensure the right connection is being used around anywhere we interact
with the `LooseForeignKeys::DeletedRecord` table. This spec was
otherwise failing when using 2 databases because the it was choosing the
wrong database to load the `LooseForeignKeys::DeletedRecord` from and
this meant that it did not find anything.