Add ability to search for an escalation policy by name
This adds a name argument to the escalationPolicies field
on projects, which allows a caller to provide a string
to search through escalation policies. Utility is limited
in the moment, as only one escalation policy is available
on each project, but that will change in future.

This change is needed to easily take advantage of shared
sidebar components in the UI.

Changelog: added
EE: true