Add EE only version of EventStore configuration
We want to be able to subscribe to events with workers that are only
declared in the EE version of GitLab. Add an EE version of the
EventStore class that we can declare these subscribers in. Also ensure
that the publish event doesn't raise an error when there are not
subscribers for that event. This stops the CE code from raising an error
when a subscriber is only declared in EE.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/351423
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79291