Add GTM enhanced ecommerce transaction event to checkout
Related to:
https://gitlab.com/gitlab-com/marketing/digital-experience/buyer-experience/-/issues/271

Add the transaction enhanced ecommerce event to new subscription page,
in the confirmOrder action with successful response.

trackTransaction will only fire when the ops feature flag for
gitlab_gtm_datalayer is enabled, and GTM is running on the page.

This commit also adds/adjusts Jest specs for the GTM tracking functions.