Show underscores in branch name in merge request
The branch name in the link to a branch on an unmerged merge request has
`overflow: hidden` set.  This means that setting the line-height setting
too small has the effect of truncating certain low descenders, like
underscore characters in certain fonts.

Removing the line height setting causes the text to move up slightly, no
longer lining up with the rest of the line, but we can fix that by
moving the margin bottom offset down a bit more to acommodate that.

Changelog: fixed