Fix editor browser modal when creating new MR
The pipeline editor will prevent user with uncommitted
changes to leave the page with a browser alert. This
was incorrectlybeing shown after a user had commited and
had checked the option to create a new MR. This fix
now ensure that the update to the internal Vue components
has occurred before checking if we should stop the user
from navigating.

Changelog: fixed