Set gitaly-backup to create incremental backups
Incremental backups require a different filesystem layout.
In `gitaly-backup` we do this using `-layout pointer`. This writes files
called `LATEST` that point to the latest full backup and the latest
increment.

Note that in the future we will likely need to use `-id` but for now we
are only using it to make tests repeatable.