Update BG migration cops for ci database worker
Update cops which detect offenses on `BackgroundMigrationWorker` to also
detect offenses on `CiDatabaseWorker`. The `CiDatabaseWorker` is a new
worker added for processing BG migrations when running with a decomposed
database. While the two workers run against separate databases, their
use should otherwise be the same, and the cops should flag the same
offenses on both.