Fix branch switcher for apollo 3 migration
The branch switcher in its current state was not
compatible with apollo 3. There were some issues
in the tests that were not supposed to occur. To
fix that, we remove the manual query execution
and instead rely solely on the smart query
execution. Since the query refire as soon
as variables are changed, we can then have
a much safer, automatic way for the smart
query to be fired whenever we need.