Remove dependency on removed LOG_KEYS constant
This moves the knowledge about which keys GitLab-rails can log
entirely into the `ApplicationContext` class.

We were using these known keys to allow deleting Sidekiq jobs from the
queue based on context information that gets added to the jobs. So we
do need to know what kind of information to accept for this.

This also adds a spec making sure we don't push anything into the
context that we haven't defined in the KNOwN_KEYS constant.