Alias wip* methods to -> draft*
Part of migrating from WIP/Work In Progress -> Draft, we want to update
the terminology we use in the code as well. This is an iterative step,
as the actual removal of every last use of wip* methods is tedious and
is something that we can either leave for new community contributors or
to fill in during idle time.

Changelog: changed