Convert timestamps to utc in license usage export
Prior to this change, the timestamps included in the license usage file
that can be exported on the subscription page did not include any
timezone information. This change will ensure that the included
timestamps are converted to utc.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79832
EE: true