Exclude pending_destruction packages when creating one
When a package is created in the Packages Registry, we can have
different checks that basically verify if there is an existing package.

This check should properly ignore `pending_destruction` packages. Those
packages are going to be removed by a background job. They should
basically be ghost packages in the Package Registry that will get
cleaned up after a while.

Changelog: changed