ci: Fix the retrying of failing specs for `-ee` jobs
The "only-failure" feature takes the intersection of failed examples
with the list of files that would run without the `--only-failures`
flag.
In our case, the problem only appeared for `-ee` jobs since the default
RSpec pattern is `spec/**/*_spec.rb`, the intersection would be empty.
Passing the pattern from the first run makes the intersection return
the "failed" examples correctly.

Also, disable Crystalball for the retry run so that the first run report
is used.

Signed-off-by: Rémy Coutable <remy@rymai.me>