Remove early_prepare_for_mergeability feature flag
The feature flag was enabled on production for more than a week
now. Looking at https://log.gprd.gitlab.net/goto/11419880-87d8-11ec-a649-b7cbb8e4f62e,
the errors that are occuring are still related to Gitaly timeout.

Those errors happens after the MR is prepared for mergeability
which is why we prepare for mergeability earlier. This way even if
the rest of the job fails, the MR will already be checked for
mergeability.

This removes the feature flag and enables the early check for all.

Changelog: changed