rake/info: Stop reporting Git version
The version of Git used to be of importance in a Rails instance. Now
that Git is almost exclusively used by Gitaly and that we have no
minimum required Git version in Rails anymore it's arguably of not much
interest anymore though. But we still report the Git version that's
installed in the system, if any. This can easily be misleading though
and be taken as a hint that Git is in fact required by Rails if no Git
version is installed and we report back an "unknown" Git version.

Remove the Git version reporting altogether to reflect Rails'
independence of Git.

Changelog: removed