system_check: Remove Git configuration check
Back when Git was still invoked by Rails we had to make sure that the
Git configuration had certain options set. One of those options was
"core.autocrlf", which was required such that no conversion of line
endings is performed when reading blobs from the repository.

Nowadays this configuration isn't required by Rails anymore: it is not
in the business of reading blobs anymore. This is now the responsibility
of Gitaly, which knows to set this option globally such that all spawned
Git commands won't ever convert line endings.

Remove the system check which checked for this option. Furthermore,
update our install instructions such that we don't ask admins to set
this config entry anymore.

Changelog: removed