system_check: Remove Git version check
We're currently requiring a Git version of at least v2.33.0, which is in
fact the same minimum required Git version as required by Gitaly. This
is not a sensible check to have though: the requirement stems back from
the days when Rails still executed Git commands to serve and modify
repos, and we don't do that nowadays anymore. In fact, Rails is in a
perfectly good shape even when no Git version is installed at all. The
CNG images serve as a point in case: they do not have Git installed.

There are some exceptions where we still do execute Git commands, which
are some of Rails tasks to set up gitlab-shell and gitlab-workhorse and
to import bare repositories. None of these require any specific version
of Git though, but would work just as well with ancient versions of Git.

Drop the Git version check completely. It's misleading and ultimately
doesn't reflect reality.

Changelog: removed